@@ -4,7 +4,7 @@
     "minimum-stability": "stable",
     "prefer-stable": true,
     "require": {
-        "php": ">=7.2.5",
+        "php": "^8.0",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "composer/package-versions-deprecated": "1.11.99.2",
@@ -14,32 +14,37 @@
         "doctrine/orm": "^2.9",
         "ramsey/uuid": "^4.1",
         "sensio/framework-extra-bundle": "^6.1",
-        "symfony/amqp-messenger": "5.3.*",
-        "symfony/console": "5.3.*",
-        "symfony/dotenv": "5.3.*",
-        "symfony/event-dispatcher": "5.3.*",
+        "symfony/amqp-messenger": "^6.0.0",
+        "symfony/console": "6.0",
+        "symfony/dotenv": "^6.0.0",
+        "symfony/event-dispatcher": "6.0",
         "symfony/flex": "^1.3.1",
-        "symfony/form": "5.3.*",
-        "symfony/framework-bundle": "5.3.*",
-        "symfony/mailer": "5.3.*",
-        "symfony/messenger": "5.3.*",
+        "symfony/form": "^6.0.0",
+        "symfony/framework-bundle": "^6.0.0",
+        "symfony/mailer": "^6.0.0",
+        "symfony/messenger": "v5.4",
         "symfony/monolog-bundle": "^3.7",
-        "symfony/proxy-manager-bridge": "5.3.*",
-        "symfony/runtime": "5.3.*",
-        "symfony/security-bundle": "5.3.*",
-        "symfony/stopwatch": "5.3.*",
-        "symfony/twig-bundle": "5.3.*",
-        "symfony/validator": "5.3.*",
-        "symfony/web-profiler-bundle": "5.3.*",
-        "symfony/yaml": "5.3.*",
-        "symfonycasts/verify-email-bundle": "^1.5"
+        "symfony/proxy-manager-bridge": "^6.0.0",
+        "symfony/runtime": "^6.0.0",
+        "symfony/security-bundle": "^6.0.0",
+        "symfony/stopwatch": "^6.0.0",
+        "symfony/twig-bundle": "^6.0.0",
+        "symfony/validator": "^6.0.0",
+        "symfony/web-profiler-bundle": "^6.0.0",
+        "symfony/yaml": "^6.0.0",
+        "symfonycasts/verify-email-bundle": "^1.9"
     },
     "config": {
         "optimize-autoloader": true,
         "preferred-install": {
             "*": "dist"
         },
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "composer/package-versions-deprecated": true,
+            "symfony/flex": true,
+            "symfony/runtime": true
+        }
     },
     "autoload": {
         "psr-4": {
@@ -66,7 +71,8 @@
         ],
         "post-update-cmd": [
             "@auto-scripts"
-        ]
+        ],
+        "php-cs-fixer": "php-cs-fixer --rules=@PSR12"
     },
     "conflict": {
         "symfony/symfony": "*"
@@ -74,12 +80,14 @@
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "5.3.*"
+            "require": "6.0"
         }
     },
     "require-dev": {
         "doctrine/doctrine-fixtures-bundle": "^3.4",
+        "friendsofphp/php-cs-fixer": "^3.4",
         "fzaninotto/faker": "^1.5",
+        "squizlabs/php_codesniffer": "^3.6",
         "symfony/maker-bundle": "^1.31"
     }
 }
