@@ -10,7 +10,7 @@ use Doctrine\Migrations\AbstractMigration;
 /**
  * Auto-generated Migration: Please modify to your needs!
  */
-final class Version20210623143813 extends AbstractMigration
+final class Version20220308184004 extends AbstractMigration
 {
     public function getDescription(): string
     {
@@ -20,15 +20,15 @@ final class Version20210623143813 extends AbstractMigration
     public function up(Schema $schema): void
     {
         // this up() migration is auto-generated, please modify it to your needs
-        $this->addSql('ALTER TABLE users ADD password VARCHAR(255) NOT NULL');
-        $this->addSql('ALTER TABLE users ADD roles JSON NOT NULL');
+        $this->addSql('ALTER TABLE users ALTER phone_number DROP NOT NULL');
+        $this->addSql('ALTER TABLE users ALTER when_convenient_receive_calls DROP NOT NULL');
     }
 
     public function down(Schema $schema): void
     {
         // this down() migration is auto-generated, please modify it to your needs
         $this->addSql('CREATE SCHEMA public');
-        $this->addSql('ALTER TABLE users DROP password');
-        $this->addSql('ALTER TABLE users DROP roles');
+        $this->addSql('ALTER TABLE users ALTER phone_number SET NOT NULL');
+        $this->addSql('ALTER TABLE users ALTER when_convenient_receive_calls SET NOT NULL');
     }
 }
